[REFACTORING] Split IMAPServerRenameMailboxTest
